NFS: Add nfs4_update_server
New function nfs4_update_server() moves an nfs_server to a different
nfs_client.  This is done as part of migration recovery.

Though it may be appealing to think of them as the same thing, a
migration recovery is not the same as following a referral.

For a referral, the client has not descended into the file system
yet: it has no nfs_server, no super block, no inodes or open state.
It is enough to simply instantiate the nfs_server and super block.

For a migration, however, we have all of those things already, and
they have to be moved to a different nfs_client.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>